requirements.txt
asgiref==3.3.4
autopep8==1.5.7
backcall==0.2.0
decorator==5.0.9
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
flake8==3.9.2
ipdb==0.13.8
ipython==7.24.0
ipython-genutils==0.2.0
jedi==0.18.0
Markdown==3.3.4
matplotlib-inline==0.1.2
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
